# Add benchmark to track byte-view construction performance