# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Part of #6034

# Rationale for this change
We know constructing short strings for byte view is slow, but how slow it is? which approach can make it faster? 

This PR allows us to benchmark potential new implementations and see if they actually make it faster.

To benchmark:
```bash
cargo bench --bench arrow_reader --features="arrow test_common experimental" "arrow_array_reader/BinaryViewArray/plain encoded, mandatory, no NULLs"
```

On my machine short string is much much slower in this (sort of) end-to-end benchmark
```
arrow_array_reader/BinaryViewArray/plain encoded, short string
                        time:   [437.05 µs 437.14 µs 437.26 µs]

arrow_array_reader/BinaryViewArray/plain encoded, mandatory, no NULLs
                        time:   [126.84 µs 127.23 µs 127.58 µs]
```
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->